+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - `countAttrValues()` for getting a deep structure with counts
 - Search history is stored as parameters only, not full urls #118
 - Enabled the `noImplicitAny` TypeScript flag for added strictness, and fixed/refactored various parts as a consequence
+- The `hitCountHtml` util function now takes the numbers as a tuple
 
 ### Fixed